split the frontend structure into public, user, and admin
