+++ frontend/src/components/public/sections/Section1.jsx
