+++ frontend/src/components/public/sections/Section2.jsx
