+++ frontend/src/components/public/sections/Section3.jsx
