+++ frontend/src/components/public/sections/Section6.jsx
