+++ frontend/src/components/public/sections/Section7.jsx
