+++ frontend/src/components/public/sections/Section8.jsx
