--- frontend/src/pages/Home.jsx
@@ -1,26 +1,0 @@
-import Section1 from "../components/sections/Section1.jsx";
-import Section2 from "../components/sections/Section2.jsx";
-import Section3 from "../components/sections/Section3.jsx";
-import Section4 from "../components/sections/Section4.jsx";
-import Section5 from "../components/sections/Section5.jsx";
-import Section6 from "../components/sections/Section6.jsx";
-import Section7 from "../components/sections/Section7.jsx";
-
-
-
-export default function Home(){
-  return(
-    <>
-      <div className="body">
-        <Section1/>
-        <Section2/>
-        <Section3/>
-        <Section4/>
-        <Section5/>
-        <Section6/>
-        <Section7/>
-        </div>
-      
-    </>
-  )
-}
