+++ frontend/src/pages/public/LoginPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import  "../styles/Login.css";
+import  "../../styles/Login.css";
 
 import { NavLink } from "react-router-dom";
