+++ frontend/src/pages/public/Sponsors.jsx
