+++ frontend/src/pages/user/Volunteer.jsx
@@ -1,6 +1,6 @@
 
-import Section4 from "../components/sections/Section4.jsx";
-import VolunteerForm from "../components/forms/VolunteerForm.jsx";
+import Section4 from "../../components/public/sections/Section4.jsx";
+import VolunteerForm from "../../components/forms/VolunteerForm.jsx";
 
 
 export default function Volunteer(){
